djangoProject3/requirements.txt
absl-py==2.1.0
accelerate==0.20.3
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
backports.zoneinfo==0.2.1
banal==1.0.6
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
coqpit==0.0.17
cycler==0.11.0
databricks-cli==0.18.0
dataset==1.6.2
datasets==2.13.2
dill==0.3.6
Django==3.2.25
django-cors-headers==4.1.0
djangorestframework==3.15.1
docker==6.1.3
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.103.2
filelock==3.12.2
Flask==2.2.5
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==0.4.6
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
huggingface-hub==0.16.4
idna==3.6
importlib-metadata==5.2.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lxml==5.1.0
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
mlflow==1.30.1
MouseInfo==0.1.3
multidict==6.0.5
multiprocess==0.70.15
mysql-connector-python==8.0.33
nltk==3.8.1
numpy==1.21.6
oauthlib==3.2.2
openpyxl==3.1.2
packaging==21.3
pandas==1.3.5
Pillow==9.5.0
pluggy==1.2.0
prometheus-client==0.17.1
prometheus-flask-exporter==0.23.0
protobuf==3.20.3
psutil==5.9.8
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pymssql==2.3.1
PyMySQL==1.1.1
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==7.4.4
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytweening==1.2.0
pytz==2022.7.1
pywin32==306
PyYAML==6.0.1
querystring-parser==1.2.4
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.4.1
SQLAlchemy==1.4.52
sqlparse==0.4.4
starlette==0.27.0
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tk==0.1.0
tokenizers==0.13.3
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.2
trainer==0.0.36
transformers==4.30.2
typing_extensions==4.7.1
unzip==1.0.0
urljoin==1.0.0
urllib3==2.0.7
uvicorn==0.22.0
waitress==2.1.2
websocket-client==1.6.1
Werkzeug==2.2.3
wordcloud==1.9.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.15.0
